# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Crabbase API",
    "description": "API for managing collections and their records",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/collections": {
      "get": {
        "tags": ["Collections"],
        "summary": "List all collections",
        "operationId": "listCollections",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of collections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Collections"],
        "summary": "Create a new collection",
        "operationId": "createCollection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              },
              "examples": {
                "allDataTypes": {
                  "summary": "Collection with all data types",
                  "description": "Example collection demonstrating every available column data type",
                  "value": {
                    "name": "products",
                    "columns": [
                      {
                        "name": "title",
                        "data_type": "PlainText",
                        "index": true,
                        "related_to": null
                      },
                      {
                        "name": "description",
                        "data_type": "RichText",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "price",
                        "data_type": "Number",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "active",
                        "data_type": "Bool",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "contact_email",
                        "data_type": "Email",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "website",
                        "data_type": "Url",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "published_at",
                        "data_type": "Datetime",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "created_at",
                        "data_type": {"AutoDatetime": "now"},
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "thumbnail",
                        "data_type": "File",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "category_id",
                        "data_type": "Relation",
                        "index": false,
                        "related_to": "categories"
                      },
                      {
                        "name": "status",
                        "data_type": "Select",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "metadata",
                        "data_type": "Json",
                        "index": false,
                        "related_to": null
                      },
                      {
                        "name": "location",
                        "data_type": "GeoPoint",
                        "index": false,
                        "related_to": null
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{name}": {
      "get": {
        "tags": ["Collections"],
        "summary": "Get a collection by name",
        "operationId": "getCollection",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Collections"],
        "summary": "Update a collection",
        "operationId": "updateCollection",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Collections"],
        "summary": "Delete a collection",
        "operationId": "deleteCollection",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{name}/truncate": {
      "post": {
        "tags": ["Collections"],
        "summary": "Truncate a collection (delete all records)",
        "operationId": "truncateCollection",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection truncated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{name}/records": {
      "get": {
        "tags": ["Records"],
        "summary": "List records in a collection",
        "operationId": "listRecords",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Records"],
        "summary": "Create a new record in a collection",
        "operationId": "createRecord",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Record created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{name}/records/{id}": {
      "get": {
        "tags": ["Records"],
        "summary": "Get a record by ID",
        "operationId": "getRecord",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Records"],
        "summary": "Update a record",
        "operationId": "updateRecord",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Record updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Records"],
        "summary": "Delete a record",
        "operationId": "deleteRecord",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record deleted"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Collection": {
        "type": "object",
        "required": ["id", "name", "fields", "indexes", "created", "updated"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique collection identifier"
          },
          "name": {
            "type": "string",
            "description": "Collection name (unique)"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Column"
            },
            "description": "Column definitions for the collection"
          },
          "indexes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Column"
            },
            "description": "Index definitions for the collection"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "description": "Collection creation timestamp"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "description": "Collection last update timestamp"
          }
        }
      },
      "CollectionListResponse": {
        "type": "object",
        "required": ["items", "total", "page", "per_page"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Collection"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of collections"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "per_page": {
            "type": "integer",
            "description": "Number of items per page"
          }
        }
      },
      "CreateCollectionRequest": {
        "type": "object",
        "required": ["name", "columns"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Collection name (must be unique)"
          },
          "columns": {
            "type": "array",
            "description": "Column definitions for the collection",
            "items": {
              "$ref": "#/components/schemas/Column"
            }
          }
        }
      },
      "UpdateCollectionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New collection name (optional)"
          },
          "columns": {
            "type": "array",
            "description": "New column definitions for the collection (optional)",
            "items": {
              "$ref": "#/components/schemas/Column"
            }
          }
        }
      },
      "Column": {
        "type": "object",
        "required": ["name", "data_type", "index"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Column name"
          },
          "data_type": {
            "$ref": "#/components/schemas/DataTypes",
            "description": "Column data type"
          },
          "index": {
            "type": "boolean",
            "description": "Whether to create an index on this column"
          }
        }
      },
      "DataTypes": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "PlainText",
              "RichText",
              "Number",
              "Bool",
              "Email",
              "Url",
              "Datetime",
              "File",
              "Relation",
              "Select",
              "Json",
              "GeoPoint"
            ]
          },
          {
            "type": "object",
            "properties": {
              "AutoDatetime": {
                "type": "string",
                "description": "Auto-datetime mode (e.g. 'now', 'created', 'updated')"
              }
            },
            "required": ["AutoDatetime"]
          }
        ],
        "description": "Available column data types"
      },
      "Record": {
        "type": "object",
        "required": ["id", "data", "created", "updated"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique record identifier"
          },
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Flexible JSON object containing record data"
          },
          "created": {
            "type": "string",
            "description": "Record creation timestamp"
          },
          "updated": {
            "type": "string",
            "description": "Record last update timestamp"
          }
        }
      },
      "RecordListResponse": {
        "type": "object",
        "required": ["items", "total", "page", "per_page"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Record"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of records"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "per_page": {
            "type": "integer",
            "description": "Number of items per page"
          }
        }
      },
      "CreateRecordRequest": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Flexible JSON object containing record data"
          }
        }
      },
      "UpdateRecordRequest": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Flexible JSON object containing updated record data"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "description": "Optional additional error details"
          }
        }
      }
    }
  }
}